extracted logic from ChoreScreen to ChoreList
